requirements.txt
SQLAlchemy==2.0.25
pyTelegramBotAPI==4.26.0
requests==2.31.0
telebot==0.0.5  
psycopg2-binary==2.9.6
python-dotenv==0.21.0